sturct align

+++ pygtpnl/structures.py
@@ -3,13 +3,13 @@ Python representations of structs used in libgtpnl
 '''
 
 from ctypes import Structure, Union
-from ctypes import c_ushort, c_int, c_uint, c_uint16, c_uint32, c_uint64
+from ctypes import c_ushort, c_int, c_uint16, c_uint32, c_uint64
 from ctypes import c_byte
 
 # libgtpnl
 class V0(Structure):
-    _fields_ = [("tid", c_uint),
-                ("flowid", c_uint)
+    _fields_ = [("tid", c_uint64),
+                ("flowid", c_uint16)
                ]
 
 # libgtpnl
